requirements.txt
###### Requirements with Version Specifiers ######
#   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
asgiref==3.3.1
astroid==2.4.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
django-simple-history
Django
entrypoints==0.3
flake8
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2
pyactiveresource==2.2.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint
pyodbc
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests
six
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==16.7.9
virtualenv-clone==0.5.3
wrapt==1.12.1
#
